Add licence to give credit to sklearn for TfidfVectorizer
Refactor files to have clean imports and clean licence
